+++ esp-hal-common/src/soc/esp32s3/peripherals.rs
@@ -20,10 +20,12 @@ pub(crate) use self::peripherals::*;
 // peripheral (no `PSRAM`, `RADIO`, etc. peripheral in the PACs), so we're
 // creating "virtual peripherals" for them.
 crate::peripherals! {
+    ADC1 <= virtual,
+    ADC2 <= virtual,
     AES <= AES,
     APB_CTRL <= APB_CTRL,
-    APB_SARADC <= APB_SARADC,
     ASSIST_DEBUG <= ASSIST_DEBUG,
+    BT <= virtual,
     DMA <= DMA,
     DS <= DS,
     EFUSE <= EFUSE,
@@ -43,6 +45,7 @@ crate::peripherals! {
     LPWR <= RTC_CNTL,
     PCNT <= PCNT,
     PERI_BACKUP <= PERI_BACKUP,
+    PSRAM <= virtual,
     MCPWM0 <= MCPWM0,
     MCPWM1 <= MCPWM1,
     RMT <= RMT,
@@ -50,7 +53,6 @@ crate::peripherals! {
     RSA <= RSA,
     RTC_I2C <= RTC_I2C,
     RTC_IO <= RTC_IO,
-    SENS <= SENS,
     SENSITIVE <= SENSITIVE,
     SHA <= SHA,
     SPI0 <= SPI0,
@@ -67,15 +69,11 @@ crate::peripherals! {
     UART2 <= UART2,
     UHCI0 <= UHCI0,
     UHCI1 <= UHCI1,
+    ULP_RISCV_CORE <= virtual,
     USB0 <= USB0,
     USB_DEVICE <= USB_DEVICE,
     USB_WRAP <= USB_WRAP,
     WCL <= WCL,
-    XTS_AES <= XTS_AES,
-
-    // Virtual peripherals:
-    BT <= virtual,
-    PSRAM <= virtual,
-    ULP_RISCV_CORE <= virtual,
     WIFI <= virtual,
+    XTS_AES <= XTS_AES,
 }