+++ esp32-hal/examples/adc.rs
@@ -6,10 +6,10 @@
 #![no_main]
 
 use esp32_hal::{
-    adc::{AdcConfig, Attenuation, ADC, ADC2},
+    adc::{AdcConfig, Attenuation, ADC},
     clock::ClockControl,
     gpio::IO,
-    peripherals::Peripherals,
+    peripherals::{Peripherals, ADC2},
     prelude::*,
     Delay,
 };
@@ -25,12 +25,10 @@ fn main() -> ! {
     let io = IO::new(peripherals.GPIO, peripherals.IO_MUX);
 
     // Create ADC instances
-    let analog = peripherals.SENS.split();
-
     let mut adc2_config = AdcConfig::new();
     let mut pin25 =
         adc2_config.enable_pin(io.pins.gpio25.into_analog(), Attenuation::Attenuation11dB);
-    let mut adc2 = ADC::<ADC2>::adc(analog.adc2, adc2_config).unwrap();
+    let mut adc2 = ADC::<ADC2>::adc(peripherals.ADC2, adc2_config).unwrap();
 
     let mut delay = Delay::new(&clocks);